refactor(ToolCard): extract styles from Tools to ToolCard CSS Module
